Addepted to create node to help avoid obstacles
This did not work due to the laser scanner data not being granular
enough.  Unable to detect small objects like the toys.  Abondoned
this approach.  Going to try to anticipate a toy is in the path
before heading to start location.

Added demo folder with commands to help run multiple commands to
assist with demos